+++ base/src/org/open2jam/render/entities/AnimatedEntity.java
@@ -29,7 +29,7 @@ public class AnimatedEntity extends Entity
 
     /** move the entity and change frame if necessary **/
     @Override
-    public void move(long delta)
+    public void move(double delta)
     {
             super.move(delta);
             sub_frame += delta * frames.getFrameSpeed();
